[protobuf] Fix parsing of oneof fields (#302)
Closes #297 

In protobuf, unions themselves are optional, so while parsing if any value fails
to parse, its default must not be considered. When the default is considered and
first value in the union is not specified, the union still gets parsed as the
default of the first value. So, in this instance:

oneof foo {
  int32 foo_int = 1;
  string foo_str = 2;
}

foo_str will never be parsed, as foo_int will have a default, i.e. 0 and so
foo_str "blah" will always wrongly parse as foo_int 0.

Co-authored-by: Akshay Mankar <itsakshaymankar@gmail.com>